refactor: InputView 내부에서 InputParser를 사용
