+++ src/main/java/janggi/board/position/StartAndGoalPosition.java
@@ -1,0 +1,19 @@
+package janggi.board.position;
+
+public class StartAndGoalPosition {
+    private final Position start;
+    private final Position goal;
+
+    public StartAndGoalPosition(Position start, Position goal) {
+        this.start = start;
+        this.goal = goal;
+    }
+
+    public Position getStart() {
+        return start;
+    }
+
+    public Position getGoal() {
+        return goal;
+    }
+}
